+++ auth/repository.go
@@ -2,7 +2,7 @@ package auth
 
 import (
 	"context"
-	"github.com/go-park-mail-ru/2021_1_kekEnd/kekEnd_main/models"
+	"github.com/go-park-mail-ru/2021_1_kekEnd/models"
 )
 
 type UserRepository interface {
