+++ auth/usecase/usecase.go
@@ -2,8 +2,8 @@ package usecase
 
 import (
 	"context"
-	"github.com/go-park-mail-ru/2021_1_kekEnd/kekEnd_main/auth"
-	"github.com/go-park-mail-ru/2021_1_kekEnd/kekEnd_main/models"
+	"github.com/go-park-mail-ru/2021_1_kekEnd/auth"
+	"github.com/go-park-mail-ru/2021_1_kekEnd/models"
 )
 
 type AuthUseCase struct {
