+++ go.mod
@@ -1,3 +1,3 @@
-module github.com/go-park-mail-ru/2021_1_kekEnd/kekEnd_main
+module github.com/go-park-mail-ru/2021_1_kekEnd
 
 go 1.15
